+++ pynetdicom/utils.py
@@ -6,223 +6,11 @@ import logging
 import sys
 import unicodedata
 
-from pydicom.datadict import (
-    tag_for_keyword, repeater_has_keyword, get_entry
-)
-from pydicom.dataset import Dataset
-from pydicom.tag import Tag
-
 from pynetdicom import _config
 
 
 LOGGER = logging.getLogger('pynetdicom.utils')
 
-
-# TODO: Move to dsutils.py instead
-class ElementPath(object):
-    """Class for parsing DICOM data elements defined using strings."""
-    def __init__(self, elem_path, parent=None):
-        """Initialise a new ElementPath.
-
-        Parameters
-        ----------
-        elem_path : str
-            The string describing the complete path of a DICOM data element.
-        parent : ElementPath or None
-            The parent of the current ElementPath (if there is one) or None.
-        """
-        self._child = None
-        self._value = None
-
-        self.path = elem_path
-        self.components = elem_path
-        self._parent = parent
-
-    @property
-    def child(self):
-        """Return the current object's child ElementPath (or None)."""
-        return self._child
-
-    @property
-    def parent(self):
-        """Return the current object's parent ElementPath (or None)."""
-        return self._parent
-
-    @property
-    def value(self):
-        """Return the value assigned to the data element."""
-        if self.parent is None:
-            return self._value
-
-        parent = self.parent
-        while parent.parent:
-            parent = parent.parent
-
-        return parent._value
-
-    @property
-    def components(self):
-        """Return the element's component str."""
-        return self._components
-
-    @components.setter
-    def components(self, value):
-        """Set the element's component str."""
-        elem_path = value
-        if '=' in value:
-            elem_path, elem_value = value.split('=', 1)
-            self._value = elem_value
-
-        self._components = elem_path.split('.')
-
-        if len(self._components) > 1:
-            self._child = ElementPath('.'.join(self._components[1:]), self)
-
-        # Parse current component
-        tag = self.tag
-        try:
-            # Try DICOM dictionary
-            self._entry = get_entry(tag)
-        except Exception as exc:
-            # Private element
-            self._entry = ('UN', '1', 'Unknown', False, 'Unknown')
-
-    @property
-    def is_private(self):
-        pass
-
-    @property
-    def is_sequence(self):
-        """Return True if the element component is a sequence."""
-        start = self.components[0].find('[')
-        end = self.components[0].find(']')
-        if start >= 0 or end >= 0:
-            if not (start >= 0 and end >= 0):
-                raise ValueError(
-                    'Invalid element component: missing item bracket'
-                )
-
-            if start + 1 == end:
-                raise ValueError(
-                    'Invalid element component: missing item number'
-                )
-
-            self._item_nr = int(self.components[0][start + 1:end])
-            return True
-
-        return False
-
-    @property
-    def item_nr(self):
-        """Return the element's sequence item number.
-
-        Returns
-        -------
-        int or None
-            TODO
-        """
-        if self.is_sequence:
-            return self._item_nr
-
-        return None
-
-    @property
-    def keyword(self):
-        """Return the element's keyword as a str."""
-        return self._entry[4]
-
-    @property
-    def tag(self):
-        """Return the element's tag as a pydicom.tag.Tag."""
-        tag = self.components[0]
-        if self.is_sequence:
-            tag = tag.split('[')[0]
-
-        # (gggg,eeee) based tag
-        if ',' in tag:
-            tag_g, tag_e = tag.split(',')
-            if '(' in tag_g:
-                tag_g = tag_g[1:]
-            if ')' in tag_e:
-                tag_e = tag_e[:4]
-
-            return Tag(tag_g, tag_e)
-
-        # Keyword based tag - private keywords not allowed
-        if repeater_has_keyword(tag):
-            raise ValueError("repeater elements must be specified using (gggg,eeee)")
-
-        tag = tag_for_keyword(tag)
-        # Test against None as 0x00000000 is a valid tag
-        if tag is not None:
-            return Tag(tag)
-
-        raise ValueError("Unknown keyword")
-
-    @property
-    def VR(self):
-        """Return the element's VR as str."""
-        return self._entry[0]
-
-    @property
-    def VM(self):
-        """Return the element's VM as str."""
-        return self._entry[1]
-
-
-def as_dataset(items):
-    """
-
-    * VR for private elements will be set to UN
-
-    Parameters
-    ----------
-    items : list of str
-        A list of strings describing the elements that are to be converted
-        to a DICOM dataset.
-
-    Returns
-    -------
-    pydicom.dataset.Dataset
-        The DICOM dataset created using the list of element descriptions.
-    """
-    def recurse(elem):
-        """
-        Parameters
-        ----------
-        ElementPath
-            A list of components describing the path to the element.
-        value
-            The value of the element.
-
-        Returns
-        -------
-        pydicom.dataset.Dataset
-            A DICOM dataset representation of `elements`.
-        """
-        ds = Dataset()
-        #print('is seq', elem.is_sequence)
-        if elem.is_sequence:
-            # Fix this -> elem is singleton
-            ds.add_new(elem.tag, 'SQ', [])
-            ds[elem.tag].value.append(recurse(elem.child))
-        else:
-            #print('set', elem, elem.tag, elem.value)
-            ds.add_new(elem.tag, elem.VR, elem.value)
-
-        return ds
-
-    ds = Dataset()
-    elements = [ElementPath(ii) for ii in items]
-    # Each element component may be:
-    #   tag based: (0010,0010), (0010,0020)[1]
-    #   keyword based: PatientName, BeamSequence[0]
-    # Can have mixed components within a single element string
-    # Can't have comma in keyword based element components
-    for elem in elements:
-        ds.update(recurse(elem))
-
-    return ds
 
 def pretty_bytes(bytestream, prefix='  ', delimiter='  ', items_per_line=16,
                  max_size=512, suffix=''):
